+++ Front-End/src/components/Menubar.tsx
@@ -1,6 +1,6 @@
 import Container from 'react-bootstrap/Container';
 import './index.css';
-
+import DownloadMap from './DownloadMap';
 
 import {Navbar, Nav, NavDropdown, Form, FormControl} from 'react-bootstrap'
 export default function Menubar({
@@ -22,7 +22,7 @@ export default function Menubar({
             
             <NavDropdown title="File" id="basic-nav-dropdown" >
               <NavDropdown.Item id="action" href="#action/3.2">Export Project </NavDropdown.Item>
-              <NavDropdown.Item id="action">Export Map</NavDropdown.Item>
+              <NavDropdown.Item id="action" onClick={DownloadMap}>Export Map</NavDropdown.Item>
               <NavDropdown.Item id="action" href='/projects/'>Exit</NavDropdown.Item>
             </NavDropdown>
