+++ Front-End/src/visualizer/index.tsx
@@ -25,7 +25,6 @@ import EditPacketModal from './modals/EditPacketModal';
 import CustomNodeData from './nodeMap/CustomNodeData';
 import HideNodesModal from './modals/HideNodesModal';
 import ProjectState from '../projects/new/ProjectState';
-import DownloadButton from '../components/DownloadButton';
 const MENU_ID = 'packet-context-menu';
 
 function Visualizer() {
@@ -467,8 +466,7 @@ function Visualizer() {
                     onPlay={onPlay}
                     />
                 </div>
-                <div className='node-map-container-content' id='download-image'>
-                    <DownloadButton/>
+                <div className='node-map-container-content'>
                     <NodeMap
                         nodes={nodes}
                         edges={edges}
